@@ -1,11 +1,13 @@
 {
   "name": "nimiq-voting-app",
+  "type": "module",
   "version": "0.1.0",
   "scripts": {
-    "serve": "vue-cli-service serve",
-    "build": "vue-cli-service build",
-    "test:unit": "vue-cli-service test:unit",
-    "lint": "vue-cli-service lint --fix"
+    "dev": "vite",
+    "build": "vue-tsc -b && vite build",
+    "preview": "vite preview",
+    "lint": "eslint .",
+    "lint:fix": "eslint . --fix"
   },
   "dependencies": {
     "@nimiq/browser-warning": "^1.1.0",
@@ -13,40 +15,22 @@
     "@nimiq/hub-api": "^1.2.3",
     "@nimiq/style": "^0.8.5",
     "@nimiq/utils": "^0.12.4",
-    "@nimiq/vue-components": "https://github.com/nimiq/vue-components.git#build/vote",
-    "@types/string-hash": "^1.1.1",
-    "@types/vuedraggable": "^2.23.0",
-    "core-js": "^3.6.4",
+    "@nimiq/vue3-components": "github:nimiq/vue3-components#build/vote",
     "distinct-colors": "^2.1.0",
     "string-hash": "^1.1.3",
-    "vue": "^2.6.11",
-    "vue-class-component": "^7.2.2",
-    "vue-property-decorator": "^8.3.0",
-    "vuedraggable": "^2.23.2"
+    "vue": "^3.5.13",
+    "vuedraggable": "^4.1.0"
   },
   "devDependencies": {
-    "@types/jest": "^24.0.19",
-    "@typescript-eslint/eslint-plugin": "^2.18.0",
-    "@typescript-eslint/parser": "^2.18.0",
-    "@vue/cli-plugin-babel": "~4.2.0",
-    "@vue/cli-plugin-eslint": "~4.2.0",
-    "@vue/cli-plugin-typescript": "~4.2.0",
-    "@vue/cli-plugin-unit-jest": "~4.2.0",
-    "@vue/cli-service": "~4.2.0",
-    "@vue/eslint-config-airbnb": "^5.0.2",
-    "@vue/eslint-config-typescript": "^5.0.1",
-    "@vue/test-utils": "1.0.0-beta.31",
-    "copy-webpack-plugin": "^5.1.1",
-    "eslint": "^6.7.2",
-    "eslint-plugin-import": "^2.20.1",
-    "eslint-plugin-vue": "^6.1.2",
-    "node-sass": "^9.0.0",
-    "pug": "^3.0.1",
-    "pug-plain-loader": "^1.0.0",
-    "sass-loader": "^8.0.2",
-    "stylus": "^0.54.7",
-    "stylus-loader": "^3.0.2",
-    "typescript": "~3.7.5",
-    "vue-template-compiler": "^2.6.11"
+    "@antfu/eslint-config": "^7.2.0",
+    "@types/string-hash": "^1.1.1",
+    "@vitejs/plugin-vue": "^5.2.0",
+    "eslint": "^9.39.2",
+    "typescript": "^5.7.0",
+    "vite": "^6.0.0",
+    "vite-plugin-static-copy": "^2.2.0",
+    "vite-plugin-top-level-await": "^1.6.0",
+    "vite-plugin-wasm": "^3.5.0",
+    "vue-tsc": "^2.2.0"
   }
 }
