@@ -5,7 +5,7 @@ export function unique<T>(array: T[], compare: (a: T, b: T) => boolean = identit
     return array.filter((e, index) => index === array.findIndex((v: T) => compare(e, v)));
 }
 
-export function allUnique <T>(array: T[], compare: (a: T, b: T) => boolean = identityCompare) {
+export function allUnique<T>(array: T[], compare: (a: T, b: T) => boolean = identityCompare) {
     return array.length === unique(array, compare).length;
 }
 
